Billing Update
For pros:
Add support for connecting Stripe account to our platform by 'approved' professionals (currently relies on an uncommitted/undeployed change to the bridge to determine if an account is 'approved').
Also allows revocation of access to a connected Stripe account.

For patients:
On opening a conversation, we now query to see if the contact has a Stripe account and products for sale, and if so, we present a menu in the action bar with the product names and prices. Tapping a product will purchase the product. Still-to-do: wire up the payment input process - currently uses a hard-coded stored test card.

Also converted all hard-coded strings to strings.xml references, removed some debugging Toasts, fixed some bugs, updated the gradle-witness dependency checksums, etc

Oh and the app now apparently exceeds the 64k method reference limit so now requires mutil-dexing. Joy.